refactor(analysis): extract near-duplicate clustering
